error handling for posts

+++ backend/controllers/post.js
@@ -9,20 +9,20 @@ exports.createPost = (req, res, next) => {
         ...postObject,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     })
-        .then(() => res.status(200).json({ message: 'post créé' }))
+        .then(() => res.status(201).json({ message: 'Post créé' }))
         .catch(error => res.status(400).json({ error }));
 }
 
 exports.getAllPosts = (req, res, next) => {
     Post.findAll({ order: [['updatedAt', 'DESC']] })
         .then(posts => res.status(200).json(posts))
-        .catch(error => res.status(400).json({ error }));
+        .catch(error => res.status(404).json({ error }));
 }
 
 exports.getOnePost = (req, res, next) => {
     Post.findOne({ where: { id: req.params.id } })
         .then(post => res.status(200).json(post))
-        .catch(error => res.status(400).json({ error }));
+        .catch(error => res.status(404).json({ error }));
 }
 
 exports.modifyPost = (req, res, next) => {
@@ -38,34 +38,45 @@ exports.modifyPost = (req, res, next) => {
 
     const updatePost = () => {
         Post.update(postUpdated, { where: { id: req.params.id } })
-            .then(() => res.status(200).json({ message: 'post modifié' }))
+            .then(() => res.status(201).json({ message: 'Post modifié' }))
             .catch(error => res.status(400).json({ error }));
     }
 
     if (req.file) {
         Post.findOne({ where: { id: req.params.id } })
             .then(post => {
-                const filename = post.imageUrl.split('/images')[1];
-                fs.unlink(`images/${filename}`, () => {
+                if (post.imageUrl) {
+                    const filename = post.imageUrl.split('/images')[1];
+                    fs.unlink(`images/${filename}`, () => {
+                        updatePost();
+                    })
+                } else {
                     updatePost();
-                })
+                }
             })
-            .catch(error => res.status(500).json({ error }));
+            .catch(error => res.status(404).json({ error }));
     } else {
         updatePost();
     }
 }
 
 exports.deletePost = (req, res, next) => {
-    Post.findOne({ where: { id: req.params.id } })
-        .then(post => {
-            const filename = post.imageUrl.split('/images')[1];
-            fs.unlink(`images/${filename}`, () => {
-                Post.destroy({ where: { id: req.params.id } })
-                    .then(() => res.status(200).json({ message: 'post supprimé' }))
-                    .catch(error => res.status(400).json({ error }));
+    if (req.file) {
+        Post.findOne({ where: { id: req.params.id } })
+            .then(post => {
+                const filename = post.imageUrl.split('/images')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Post.destroy({ where: { id: req.params.id } })
+                        .then(() => res.status(200).json({ message: 'Post supprimé' }))
+                        .catch(error => res.status(400).json({ error }));
+                })
             })
-        })
+            .catch(error => res.status(404).json({ error }));
+    } else {
+        Post.destroy({ where: { id: req.params.id } })
+            .then(() => res.status(200).json({ message: 'Post supprimé' }))
+            .catch(error => res.status(400).json({ error }));
+    }
 }
 
 exports.sendLike = (req, res, next) => {
@@ -81,9 +92,9 @@ exports.sendLike = (req, res, next) => {
                     postId: req.params.id,
                     userId: req.userId
                 })
-                    .then(() => res.status(200).json({ message: 'post Liké' }))
+                    .then(() => res.status(201).json({ message: 'Post liké' }))
                     .catch(error => res.status(400).json({ error }));
             }
         })
-        .catch(error => res.status(400).json({ error }));
+        .catch(error => res.status(404).json({ error }));
 }
